# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zynk API
  description: |
    End-to-end encrypted messaging platform API.
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the `Authorization` header.
    Tokens are obtained via `/auth/login` or `/auth/register`.
    
    ## Rate Limiting
    - Global: 100 requests / 15 minutes per IP
    - Login: 5 attempts / 15 minutes per IP
    - Registration: 3 attempts / 15 minutes per IP
    - Messages: 60 per minute per user
    - File uploads: 20 per hour per user
    
    ## Encryption
    All message content is end-to-end encrypted using the Signal Protocol.
    The server only stores encrypted ciphertext.
  version: 1.0.0
  contact:
    name: Zynk Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        conversation_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        member_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
paths:
...
  /groups:
    post:
      tags: [Groups]
      summary: Create a group
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                description: { type: string, maxLength: 500 }
                avatar_url: { type: string, format: uri, nullable: true }
                member_ids:
                  type: array
                  items: { type: string, format: uuid }
                  minItems: 1
                  maxItems: 31
              required: [name, member_ids]
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}:
    get:
      tags: [Groups]
      summary: Get group details
      operationId: getGroup
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Group details
    put:
      tags: [Groups]
      summary: Update group info
      operationId: updateGroup
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                avatar_url: { type: string }
      responses:
        '200':
          description: Group updated
  /groups/{groupId}/members:
    post:
      tags: [Groups]
      summary: Add members to group
      operationId: addGroupMembers
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                member_ids:
                  type: array
                  items: { type: string, format: uuid }
              required: [member_ids]
      responses:
        '200':
          description: Members added

  # ======================== Files ========================
...